Create README.md
My name is Lucas Mwita. I have been in Healthcare for quite a while and now I'm trying to make a transfer to Fintech. The reason behind these changes is because  technology is changing the way we work, jobs opportunities, etc so basically I want to plug in and position myself and I think Fintech will be the vehicle to take me there,